+++ wezterm-gui/src/termwindow/wallpaper.rs
@@ -487,7 +487,7 @@ impl WallpaperOverlay {
             .into();
 
         let top_bar_height = if term_window.show_tab_bar && !term_window.config.tab_bar_at_bottom {
-            term_window.tab_bar_pixel_height().unwrap()
+            term_window.tab_bar_pixel_height_lossy()
         } else {
             0.
         };